Organise class into package
