+++ src/main/java/com/Commands/ListCommand.java
@@ -1,0 +1,14 @@
+package com.Commands;
+
+import com.Nimbus.Storage;
+import com.Nimbus.Task;
+import com.Nimbus.Ui;
+import java.util.ArrayList;
+
+public class ListCommand extends Command {
+
+    @Override
+    public void execute(Ui ui, Storage storage, ArrayList<Task> tasks) {
+        return;
+    }
+}
