+++ src/main/java/com/Nimbus/Deadline.java
@@ -1,9 +1,10 @@
-import java.time.DateTimeException;
+package com.Nimbus;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
     protected String deadline;
 
     Deadline(String description, String deadline) {
