+++ src/main/java/com/Nimbus/Event.java
@@ -1,5 +1,6 @@
+package com.Nimbus;
+
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
