+++ src/main/java/com/Nimbus/InvalidArgumentException.java
@@ -1,3 +1,5 @@
+package com.Nimbus;
+
 public class InvalidArgumentException extends Exception{
     public InvalidArgumentException(String msg) {
         super(msg);
