+++ src/main/java/com/Nimbus/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package com.Nimbus;
+
 public class InvalidCommandException extends Exception{
     public InvalidCommandException(String msg) {
         super(msg + ": Is not a valid command.");
